pyproject.toml
[project]
name = "ourhour"
version = "0.1.0"
description = ""
authors = [
    {name = "benchprincess",email = "realjiwon98@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "drf-spectacular (>=0.27.0,<1.0.0)",
    "django-filter (>=24.0,<25.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-debug-toolbar (>=6.1.0,<7.0.0)",
    "mysqlclient (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
isort = "^7.0.0"
mypy = "^1.19.0"
django-stubs = "^5.2.8"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
coverage = "^7.13.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = false

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "*.tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "*.views.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "*.serializers.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.coverage.report]
fail_under = 80

[tool.coverage.run]
omit = [
  ".mypy_cache/**/*",
  "*/migrations/*",
  "*/*test*.py",
]